Create .DS_Store
I was able to read in the geojson file, I cannot read in any of the csv files - R is giving an error that there are "more columns than column names" which is strange because I checked and that does not appear to be the case in any of the csv files. Do either of you see the problem?